# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Batch notification management and async processing API.",
        "title": "Notification API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/notifications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status (pending|processing|success|failure|cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Channel (email|sms|push)",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 start time",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 end time",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notification.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/batch": {
            "post": {
                "description": "Creates up to 1000 notifications in a single request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create batch notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Correlation ID",
                        "name": "X-Correlation-ID",
                        "in": "header"
                    },
                    {
                        "description": "Batch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notification.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/notification.BatchCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/batch/{batchId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notifications by batch ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "batchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notification.NotificationResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/cancel": {
            "patch": {
                "description": "Cancels every notification currently in pending status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Cancel all pending notifications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notification.CancelPendingResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notification.NotificationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "notification.BatchCreateResponse": {
            "type": "object",
            "properties": {
                "batchId": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
        "notification.CancelPendingResponse": {
            "type": "object",
            "properties": {
                "cancelledCount": {
                    "type": "integer"
                }
            }
        },
        "notification.CreateNotificationRequest": {
            "type": "object",
            "required": [
                "notifications"
            ],
            "properties": {
                "notifications": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/notification.NotificationRequest"
                    }
                }
            }
        },
        "notification.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "notification.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notification.NotificationResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "notification.NotificationRequest": {
            "type": "object",
            "required": [
                "channel",
                "content",
                "priority",
                "recipient"
            ],
            "properties": {
                "channel": {
                    "type": "string",
                    "enum": [
                        "email",
                        "sms",
                        "push"
                    ]
                },
                "content": {
                    "type": "string"
                },
                "priority": {
                    "type": "string",
                    "enum": [
                        "high",
                        "medium",
                        "low"
                    ]
                },
                "recipient": {
                    "type": "string"
                }
            }
        },
        "notification.NotificationResponse": {
            "type": "object",
            "properties": {
                "batchId": {
                    "type": "string"
                },
                "cancelledAt": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}